Module iocutil::prelude

Re-exports

pub use crate::alienvault::hashes_in;

pub use crate::alienvault::AlienVaultOTXClient;

pub use crate::alienvault::Pulse;

pub use crate::alienvault::QueryType;

pub use crate::contenthash::remove_overlap;

pub use crate::contenthash::ContentHash;

pub use crate::datetime::days_ago;

pub use crate::datetime::vtdatetime;

pub use crate::hasher::Hasher;

pub use crate::sample;

pub use crate::virustotal::scan_id;

pub use crate::virustotal::VirusTotalClient;

pub use crate::SampleHash;

pub use crate::at;

pub use crate::day;

pub use crate::fs;

pub use crate::la;

pub use crate::ls;

pub use crate::p;

Structs

DateTime
ISO 8601 combined date and time with time zone.

Duration
ISO 8601 time duration with nanosecond precision. This also allows for the negative duration; see individual methods for details.

Utc
The UTC time zone. This is the most efficient time zone when you don't need the local time. It is also used as an offset (which is also a dummy type).

Traits

TryFrom
Simple and safe type conversions that may fail in a controlled way under some circumstances. It is the reciprocal of TryInto.

TryInto
An attempted conversion that consumes self, which may or may not be expensive.